+++ opencv/src/gen/java/org/bytedeco/opencv/opencv_core/MatVectorVector.java
@@ -6,6 +6,9 @@ import java.nio.*;
 import org.bytedeco.javacpp.*;
 import org.bytedeco.javacpp.annotation.*;
 
+import static org.bytedeco.openblas.global.openblas_nolapack.*;
+import static org.bytedeco.openblas.global.openblas.*;
+
 import static org.bytedeco.opencv.global.opencv_core.*;
 
 @Name("std::vector<std::vector<cv::Mat> >") @Properties(inherit = org.bytedeco.opencv.presets.opencv_core.class)